Add DriverFormAndInputs component for driver details management
