+++ src/components/molecules/tabs/logisticsForms/driverForm/components/DriverFormAndInputs/index.ts
@@ -1,0 +1,2 @@
+export { DriverFormAndInputs } from "./DriverFormAndInputs";
+export type { DriverFormAndInputsProps } from "./DriverFormAndInputs";
